National correspondents can't edit reviewers
